+++ src/app/[locale]/layout.tsx
@@ -2,8 +2,8 @@ import { Container, Stack } from '@mui/material';
 import { NextIntlClientProvider } from 'next-intl';
 import { getMessages } from 'next-intl/server';
 
-import Footer from '@/widgets/footer/Footer';
-import Header from '@/widgets/header/Header';
+import Footer from '@/widgets/footer/footer.tsx';
+import Header from '@/widgets/header/header.tsx';
 
 export default async function LocaleLayout({
   children,
