+++ src/widgets/footer/footer.tsx
